[TIR] Added ApplyPhysicalLayout as part of StorageFlatten.
For any buffers that have `BufferNode::physical_layout` defined,
rewrite access into the buffer such that they use the updated
ordering.

+++ python/tvm/te/schedule.py
@@ -558,7 +558,7 @@ class Stage(Object):
             )
 
         var_names = list(params)
-        logical_index = [tvm.tir.Var(name, dtype="int64") for name in var_names]
+        logical_index = [tvm.tir.Var(name, dtype="int32") for name in var_names]
         physical_index = mapping_function(**dict(zip(var_names, logical_index)))
 
         _ffi_api.StageSetPhysicalLayout(self, logical_index, physical_index)
